:rocket: Start reading Hands-On Functional Programming with Typescript
Loving functional programming, starting to appreciate having to learn Racket in my first year CS courses. Applying the concepts of functional programming to the code I produce at work will hopefully make it less of a hassle for future devs to debug, and make it easier for me to read and maintain the code I write going forwards.