:rocket: Start Fluent Python 2nd Edition
Add folders for each chapter. “Python is an easy to learn, powerful programming language.” Those are the first words of the official Python 3.10 Tutorial. That is true, but there is a catch: because the language is easy to learn and put to use, many practicing Python programmers leverage only a fraction of its powerful features.