:tada: TDD with doctest is segoi 😍
Luciano Ramalho is the goat for teaching python. He wrote: "Firmly establishing what is to be done before thinking about how to do it helps focus our coding effort. Writing tests first is the basis of test driven development (TDD) and I’ve also found it helpful when teaching. If you are unfamiliar with doctest"